requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu  #
altair
tiktoken
langchain-community
together 
chromadb
langchain-huggingface
langchain-chroma
pypdf
keyboard

#  huggingface llms
huggingface-hub  #

# instructor embeddings 
InstructorEmbedding  #
sentence-transformers  #